--- Ghala/Views/DispatchView.swift
@@ -1,20 +1,0 @@
-//
-//  DispatchView.swift
-//  Ghala
-//
-//  Created by mroot on 19/05/2022.
-//
-
-import SwiftUI
-
-struct DispatchView: View {
-    var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
-    }
-}
-
-struct DispatchView_Previews: PreviewProvider {
-    static var previews: some View {
-        DispatchView()
-    }
-}
